Tutorial on advanced topics regarding functions
Not quite finished: some of the function tests within main aren't yet implemented, but you're welcome to write those tests yourself before I finish the tutorial. Mostly all of the explanations are there for creating the functions, though I might still be missing an important comment or two on how to use some of the last few functions.