Example of Creating a Math Library in C for C++
The main.cpp file is the driver to test the library, to use the library you only need to put QMath.h and QMath.c into your project and #include "QMath.h" at the top of the file where you want to use its functions. Feel free to add functions to the library, just remember that 1. the bool data type doesn't exist in C, so use char or short instead, and 2. structs don't work as you'd expect in C, so avoid if possible.